+++ lib/src/constants.dart
@@ -1,11 +1,7 @@
 part of sdk;
 
-var _confirmationRequirement = 0;
-
-/**
- * !IMPORTANT!
- * DO NOT USE THESE PRIVATE KEYS IN PRODUCTION
- */
+/// !IMPORTANT!
+/// DO NOT USE THESE PRIVATE KEYS IN PRODUCTION
 const _predefinedPrivateKeys = [
   "0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d",
   "0x6cbed15c793ce57650b9877cf6fa156fbef513c4e6134f022a85b1ffdd59b2a1",