+++ lib/src/mock-ethereum.dart
@@ -22,6 +22,11 @@ class MockEthereum {
     }
     this._ganacheProcess = await Process.start('ganache-cli', ganacheOptions);
     this._port = port;
+    this._ganacheProcess.exitCode.then((value) {
+      if (value != 0) {
+        throw new Exception("ganache exit with code $value");
+      }
+    });
     sleep(Duration(seconds: 3));
   }
 
@@ -29,6 +34,7 @@ class MockEthereum {
     if (this._ganacheProcess == null) {
       return Future.value(0);
     }
+    sleep(Duration(seconds: 1));
     this._ganacheProcess.kill(ProcessSignal.sigint);
     var exitCode = this._ganacheProcess.exitCode;
     this._ganacheProcess = null;
@@ -37,6 +43,31 @@ class MockEthereum {
   }
 
   String get endpoint => "http://localhost:${this._port}";
+
+  Future<List<Account>> predefinedAccounts() async {
+    return await Future.wait(
+        _predefinedPrivateKeys.map((key) => Account.retrieve(key)));
+  }
+
+  Future<void> mineBlock() async {
+    var data = {
+      "id": 1337,
+      "jsonrpc": "2.0",
+      "method": "evm_mine",
+      "params": [1231006505000]
+    };
+    var response =
+        await post("http://localhost:8545", body: jsonEncode(data), headers: {
+      'Content-Type': "application/json",
+    });
+    if (response.statusCode != 200) {
+      throw Future.error("http status code is not 200");
+    }
+    var result = jsonDecode(response.body);
+    if (result['error'] != null) {
+      throw Future.error(result['error']);
+    }
+  }
 }
 
 // void main() async {