+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 First version of jasmine_eth. 
 This version is not tested and should not be used. 
+
+# 0.0.2
+
+Tested the package. 
+Add a usage example.